@@ -1,14 +1,14 @@
 [metadata]
-name = ImSwitch
+name = ImSwitchUC2
 version = attr: imswitch.__version__
-author = Xavier Casas Moreno
-author_email = xaviercm@kth.se
+author = Benedict Diederich (openUC2)
+author_email = benedictdied@gmail.com
 description = Microscopy control
 long_description = file: README.md
 long_description_content_type = text/markdown
-url = https://github.com/kasasxav/ImSwitch
+url = https://github.com/openUC2/ImSwitch
 project_urls =
-    Bug Tracker = https://github.com/kasasxav/ImSwitch/issues
+    Bug Tracker = https://github.com/openUC2/ImSwitch/issues
 classifiers =
     Programming Language :: Python :: 3
     License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)
@@ -17,7 +17,7 @@ classifiers =
 [options]
 packages = find:
 include_package_data = True
-python_requires = >=3.7
+python_requires = >=3.8
 install_requires =
     coloredlogs >= 15
     colour-science >= 0.3
@@ -32,7 +32,7 @@ install_requires =
     #psutil >= 5.4.8s
     PyQt5 >= 5.15.2
     pyqtgraph >= 0.12.1
-    microscope
+    microscope >= 0.3.0
     pyserial >= 3.4
     QDarkStyle >= 3
     QScintilla >= 2.12
@@ -41,23 +41,30 @@ install_requires =
     scikit-image >= 0.18
     Send2Trash >= 1.8
     tifffile >= 2020.11.26
-    pulsestreamer
-    zeroconf
+    zeroconf >= 0.39.0
     pymba >= 0.3.7
-    opencv-python
-    Pyro5
-    useq-schema
-    pco
-    visa
-    napari-sim-processor
+    opencv-python >= 4.6.0.66
+    pco >= 0.1.0
+    #napari-sim-processor
+    UC2-REST >=0.1.0.11
+    esptool>=4.2.1
+    setuptools<60.0.0
+    pypylon>=1.8.0
+    # optional:
+    #openflexure-microscope-client >=0.1.6
     #NanoImagingPack @ https://gitlab.com/bionanoimaging/nanoimagingpack/
-    napari_sim_processor
+    #napari_sim_processor
     #git+https://github.com/alliedvision/VimbaPython
-
     #NanoImagingPack @ git+ssh://git@gitlab.com/bionanoimaging/nanoimagingpack
     #git+https://github.com/alliedvision/VimbaPython
-    #git+https://github.com/tlambert03/useq-schema
     #https://stackoverflow.com/questions/58612306/how-to-fix-importerror-dll-load-failed-while-importing-win32api
+    pywin32 >= 304
+    ome_zarr >= 0.6.1
+    Pyro5 >= 5.14
+    useq-schema >= 0.1.3
+    fastAPI >= 0.86.0
+    uvicorn[standard] >= 0.19.0
+
 [options.entry_points]
 console_scripts =
     imswitch = imswitch.__main__:main
