@@ -2,7 +2,8 @@ from .BetterPushButton import BetterPushButton
 from .BetterSlider import BetterSlider
 from .CheckableComboBox import CheckableComboBox
 from .FloatSlider import FloatSlider
-from .dialogtools import askYesNoQuestion, askForFilePath, askForFolderPath, askForTextInput
+from .dialogtools import askYesNoQuestion, askForFilePath, askForFolderPath, askForTextInput, informationDisplay
 from .imagetools import bestLevels, minmaxLevels
 from .stylesheet import getBaseStyleSheet
 from .texttools import ordinalSuffix
+from .FileWatcher import FileWatcher
