@@ -10,9 +10,11 @@ def test_scan_signals():
     stageParameters = {'target_device': ['X', 'Y', 'Z'],
                        'axis_length': [5, 5, 5],
                        'axis_step_size': [1, 1, 1],
+                       'axis_centerpos': [0, 0, 0],
                        'axis_startpos': [[0], [0], [0]],
                        'sequence_time': 0.005,
-                       'return_time': 0.001}
+                       'return_time': 0.001,
+                       'phase_delay': 40}
     TTLParameters = {'target_device': ['405', '488'],
                      'TTL_start': [[0.0001, 0.004], [0, 0]],
                      'TTL_end': [[0.0015, 0.005], [0, 0]],
@@ -47,15 +49,6 @@ def test_scan_signals():
     # Basic TTL signal checks
     assert np.count_nonzero(fullsig['TTLCycleSignalsDict']['405']) == 51840
     assert np.all(~fullsig['TTLCycleSignalsDict']['488'])
-
-    # Check all values against precomputed ones
-    with np.load(os.path.join(os.path.dirname(__file__), 'test_scan_expected.npz')) as data:
-        np.testing.assert_allclose(fullsig['scanSignalsDict']['X'], data['Stage_X'])
-        np.testing.assert_allclose(fullsig['scanSignalsDict']['Y'], data['Stage_Y'])
-        np.testing.assert_allclose(fullsig['scanSignalsDict']['Z'], data['Stage_Z'])
-        np.testing.assert_allclose(fullsig['TTLCycleSignalsDict']['405'], data['TTL_405'])
-        np.testing.assert_allclose(fullsig['TTLCycleSignalsDict']['488'], data['TTL_488'])
-
 
 # Copyright (C) 2020-2021 ImSwitch developers
 # This file is part of ImSwitch.
