@@ -6,12 +6,15 @@ from imswitch.imcommon.controller import MainController, PickDatasetsController
 from imswitch.imcommon.model import (
     ostools, initLogger, generateAPI, generateShortcuts, SharedAttributes
 )
+from imswitch.imcommon.framework import Thread
+from .server import ImSwitchServer
 from imswitch.imcontrol.model import configfiletools
 from imswitch.imcontrol.view import guitools
 from . import controllers
 from .CommunicationChannel import CommunicationChannel
 from .MasterController import MasterController
 from .PickSetupController import PickSetupController
+from .PickUC2BoardConfigController import PickUC2BoardConfigController
 from .basecontrollers import ImConWidgetControllerFactory
 
 
@@ -23,27 +26,31 @@ class ImConMainController(MainController):
         self.__options = options
         self.__setupInfo = setupInfo
         self.__mainView = mainView
-        self.__moduleCommChannel = moduleCommChannel
+        self._moduleCommChannel = moduleCommChannel
 
         # Connect view signals
         self.__mainView.sigLoadParamsFromHDF5.connect(self.loadParamsFromHDF5)
         self.__mainView.sigPickSetup.connect(self.pickSetup)
+        self.__mainView.sigPickConfig.connect(self.pickUC2Config)
         self.__mainView.sigClosing.connect(self.closeEvent)
 
         # Init communication channel and master controller
         self.__commChannel = CommunicationChannel(self, self.__setupInfo)
         self.__masterController = MasterController(self.__setupInfo, self.__commChannel,
-                                                   self.__moduleCommChannel)
+                                                   self._moduleCommChannel)
 
         # List of Controllers for the GUI Widgets
         self.__factory = ImConWidgetControllerFactory(
-            self.__setupInfo, self.__masterController, self.__commChannel, self.__moduleCommChannel
+            self.__setupInfo, self.__masterController, self.__commChannel, self._moduleCommChannel
         )
         self.pickSetupController = self.__factory.createController(
             PickSetupController, self.__mainView.pickSetupDialog
         )
         self.pickDatasetsController = self.__factory.createController(
             PickDatasetsController, self.__mainView.pickDatasetsDialog
+        )
+        self.PickUC2BoardConfigController = self.__factory.createController(
+            PickUC2BoardConfigController, self.__mainView.PickUC2BoardConfigDialog
         )
 
         self.controllers = {}
@@ -63,12 +70,20 @@ class ImConMainController(MainController):
                                                   f' is not included in your currently active'
                                                   f' hardware setup file.'
         )
-
         # Generate Shorcuts
         self.__shortcuts = None
         shorcutObjs = list(self.__mainView.widgets.values())
         self.__shortcuts = generateShortcuts(shorcutObjs)
         self.__mainView.addShortcuts(self.__shortcuts)
+
+        if setupInfo.pyroServerInfo.active:
+            self._serverWorker = ImSwitchServer(self.__api, setupInfo)
+            self.__logger.debug(self.__api)
+            self._thread = Thread()
+            self._serverWorker.moveToThread(self._thread)
+            self._thread.started.connect(self._serverWorker.run)
+            self._thread.finished.connect(self._serverWorker.stop)
+            self._thread.start()
 
     @property
     def api(self):
@@ -135,7 +150,22 @@ class ImConMainController(MainController):
         self.__factory.closeAllCreatedControllers()
         self.__masterController.closeEvent()
 
+    def pickUC2Config(self):
+        """ Let the user change which UC2 Board config is used. """
 
+        options, _ = configfiletools.loadUC2BoardConfigs()
+
+        self.pickSetupController.setSetups(configfiletools.getBoardConfigList())
+        self.pickSetupController.setSelectedSetup(options.setupFileName)
+        if not self.__mainView.showPickSetupDialogBlocking():
+            return
+        setupFileName = self.pickSetupController.getSelectedSetup()
+        if not setupFileName:
+            return
+
+        guitools.informationDisplay(self.__mainView, "Now select 'load from file' in the UC2 Config Widget and flash the pin-configuration")
+        
+        
 # Copyright (C) 2020-2021 ImSwitch developers
 # This file is part of ImSwitch.
 #
