@@ -3,75 +3,74 @@ from functools import partial
 from qtpy import QtCore, QtWidgets
 import numpy as np
 
+
 from imswitch.imcommon.model import APIExport
 from ..basecontrollers import ImConWidgetController
 from imswitch.imcontrol.view import guitools as guitools
 from imswitch.imcommon.model import initLogger, APIExport
 
+
 class LEDMatrixController(ImConWidgetController):
     """ Linked to LEDMatrixWidget."""
 
-    def __init__(self, nLedsX = 8, nLedsY=8, *args, **kwargs):
+    def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.__logger = initLogger(self)
-        
-        self.nLedsX = nLedsX
-        self.nLedsY = nLedsY
-        
+
+        # TODO: This must be easier?
+        self.nLedsX = self._master.LEDMatrixsManager._subManagers['ESP32 LEDMatrix'].Nx
+        self.nLedsY = self._master.LEDMatrixsManager._subManagers['ESP32 LEDMatrix'].Ny
+
         self._ledmatrixMode = ""
 
         # get the name that looks like an LED Matrix
-        self.ledmatrix_name = self._master.LEDMatrixsManager.getAllDeviceNames()[0]
-        self.ledmatrix = self._master.LEDMatrixsManager[self.ledmatrix_name]
-        
-        # initialize the LEDMatrix device that holds all necessary states^
-        self.LEDMatrixDevice = LEDMatrixDevice(self.ledmatrix,Nx=self.nLedsX,Ny=self.nLedsY)
-        
+        self.ledMatrix_name = self._master.LEDMatrixsManager.getAllDeviceNames()[0]
+        self.ledMatrix = self._master.LEDMatrixsManager[self.ledMatrix_name]
+
         # set up GUI and "wire" buttons
         self._widget.add_matrix_view(self.nLedsX, self.nLedsY)
         self.connect_leds()
-        
-        self._widget.ButtonAllOn.clicked.connect(self.setLEDAllOn)      
-        self._widget.ButtonAllOff.clicked.connect(self.setLEDAllOff)      
-        self._widget.ButtonSubmit.clicked.connect(self.submitLEDPattern)
-        self._widget.ButtonToggle.clicked.connect(self.toggleLEDPattern)
+
+        # initialize matrix
+        self.setAllLEDOff()
+
+        self._widget.ButtonAllOn.clicked.connect(self.setAllLEDOn)
+        self._widget.ButtonAllOff.clicked.connect(self.setAllLEDOff)
         self._widget.slider.valueChanged.connect(self.setIntensity)
-        
-        
-    @APIExport()
-    def setLEDAllOn(self):
-        self._ledmatrixMode = "allon"
-        self.LEDMatrixDevice.setAllOn()
-        for coords, btn in self._widget.leds.items():
-            if isinstance(btn, guitools.BetterPushButton):
-                btn.setChecked(True)
-                
-    @APIExport()
-    def setLEDAllOff(self):
-        self._ledmatrixMode = "alloff"
-        self.LEDMatrixDevice.setAllOff()
-        for coords, btn in self._widget.leds.items():
-            if isinstance(btn, guitools.BetterPushButton):
-                btn.setChecked(False)
 
     @APIExport()
-    def submitLEDPattern(self):
-        pass #  self.LEDMatrixDevice.setPattern()
+    def setAllLEDOn(self):
+        self.setAllLED(state=(1,1,1))
 
     @APIExport()
-    def toggleLEDPattern(self):
-        pass #self.LEDMatrixDevice.toggleLEDPattern()
-        
+    def setAllLEDOff(self):
+        self.setAllLED(state=(0,0,0))
+
+    @APIExport()
+    def setAllLED(self, state=None, intensity=None):
+        if intensity is not None:
+            self.setIntensity(intensity=intensity)
+        self.ledMatrix.setAll(state=state)
+        for coords, btn in self._widget.leds.items():
+            if isinstance(btn, guitools.BetterPushButton):
+                btn.setChecked(np.sum(state)>0)
+
     @APIExport()
     def setIntensity(self, intensity=None):
         if intensity is None:
             intensity = int(self._widget.slider.value()//1)
-        self.LEDMatrixDevice.setIntensity(intensity=intensity)
-        
+        else:
+            # this is only if the GUI/API is calling this function
+            intensity = int(intensity)
+
+        self.ledMatrix.setLEDIntensity(intensity=(intensity,intensity,intensity))
+
     @APIExport()
-    def switchLED(self, LEDid):
+    def setLED(self, LEDid, state=None):
         self._ledmatrixMode = "single"
-        self.LEDMatrixDevice.switchLED(LEDid)
+        self.ledMatrix.setLEDSingle(indexled=int(LEDid), state=state)
+        pattern = self.ledMatrix.getPattern()
+        self._widget.leds[str(LEDid)].setChecked(np.mean(pattern.reshape(np.array(pattern.shape[0:2]).prod(),pattern.shape[-1]))>0)
 
     def connect_leds(self):
         """Connect leds (Buttons) to the Sample Pop-Up Method"""
@@ -79,58 +78,9 @@ class LEDMatrixController(ImConWidgetController):
         for coords, btn in self._widget.leds.items():
             # Connect signals
             if isinstance(btn, guitools.BetterPushButton):
-                btn.clicked.connect(partial(self.switchLED, coords))
-                
+                btn.clicked.connect(partial(self.setLED, coords))
 
-class LEDMatrixDevice():
-    
-    def __init__(self, ledMatrix, Nx=8, Ny=8):
-        self.Nx=Nx
-        self.Ny=Ny
-        self.ledMatrix = ledMatrix 
-        self.pattern = np.zeros((self.Nx*self.Ny,3))
-        self.intensity = (255,255,255)
-        self.state=None
-        
-        # Turn off LEDs
-        self.ledMatrix.setAll(intensity=(0,0,0))
-        
-    def setIntensity(self, intensity=None):
-        if intensity is None:
-            intensity = self.intensity  
-        self.pattern = (self.pattern>0)*(intensity,intensity,intensity)
-        self.intensity = (intensity,intensity,intensity)
-        self.ledMatrix.setPattern(self.pattern)
-        
-    def setPattern(self, pattern):
-        self.pattern = pattern
-        self.ledMatrix.setPattern(self.pattern)
-        self.state="pattern"
-          
-    def switchLED(self, index, intensity=None):
-        if intensity is None:
-            intensity = self.intensity
-        index = int(index)
-        if np.sum(self.pattern[index,:]):
-            self.pattern[index,:] = (0,0,0) 
-        else:
-            self.pattern[index,:] = intensity
-        self.ledMatrix.setLEDSingle(indexled=index, intensity=self.pattern[index,:])
-    
-    def setAllOn(self, intensity=None):
-        if intensity is None:
-            intensity = self.intensity
-        self.pattern = np.ones((self.Nx*self.Ny, 3))*intensity
-        self.ledMatrix.setAll(intensity=intensity)
-    
-    def setAllOff(self):
-        self.pattern = np.zeros((self.Nx*self.Ny, 3))
-        self.ledMatrix.setAll(intensity=(0,0,0))
-    
-    def toggleLEDPattern(self):
-        pass  
-        
-        
+
 # Copyright (C) 2020-2021 ImSwitch developers
 # This file is part of ImSwitch.
 #
