@@ -1,6 +1,7 @@
 import os
 import time
 from typing import Optional, Union, List
+import numpy as np
 
 from imswitch.imcommon.framework import Timer
 from imswitch.imcommon.model import ostools, APIExport
@@ -80,7 +81,7 @@ class RecordingController(ImConWidgetController):
         if saveMode == SaveMode.RAM:
             self._widget.setsaveFormat(SaveFormat.TIFF.value)
 
-    def snap(self):
+    def snap(self, name=None):
         """ Take a snap and save it to a file. """
         self.updateRecAttrs(isSnapping=True)
 
@@ -90,17 +91,19 @@ class RecordingController(ImConWidgetController):
         time.sleep(0.01)
 
         detectorNames = self.getDetectorNamesToCapture()
-        savename = os.path.join(folder, self.getFileName()) + '_snap'
+        if name is None:
+            name = '_snap'
+        savename = os.path.join(folder, self.getFileName()) + name
 
         attrs = {detectorName: self._commChannel.sharedAttrs.getHDF5Attributes()
                  for detectorName in detectorNames}
-        
+
         self._master.recordingManager.snap(detectorNames,
                                            savename,
                                            SaveMode(self._widget.getSnapSaveMode()),
                                            SaveFormat(self._widget.getsaveFormat()),
                                            attrs)
-        
+
     def snapNumpy(self):
         self.updateRecAttrs(isSnapping=True)
         detectorNames = self.getDetectorNamesToCapture()
@@ -370,10 +373,19 @@ class RecordingController(ImConWidgetController):
     def getTimelapseFreq(self):
         return self._widget.getTimelapseFreq()
 
+    
+    '''
+    def snapImage(self, name=None) -> None:
+        self.snap(name)
+    '''
     @APIExport(runOnUIThread=True)
-    def snapImage(self) -> None:
+    def snapImage(self, output: bool = False) -> Optional[np.ndarray]:
         """ Take a snap and save it to a .tiff file at the set file path. """
-        self.snap()
+        if output:
+            return self.snapNumpy()
+        else:
+            self.snap()
+
 
     @APIExport(runOnUIThread=True)
     def startRecording(self) -> None:
