@@ -1,10 +1,11 @@
 from imswitch.imcommon.model import initLogger
 from .LEDMatrixManager import LEDMatrixManager
-from imswitch.imcontrol.model.interfaces.ESP32Client import ESP32Client
 import numpy as np
 
+import uc2rest
+
 class ESP32LEDMatrixManager(LEDMatrixManager):
-    """ LEDMatrixManager for controlling LEDs and LEDMatrixs connected to an 
+    """ LEDMatrixManager for controlling LEDs and LEDMatrixs connected to an
     ESP32 exposing a REST API
     Each LEDMatrixManager instance controls one LED.
 
@@ -19,44 +20,53 @@ class ESP32LEDMatrixManager(LEDMatrixManager):
         self.__logger = initLogger(self, instanceName=name)
         self.power = 0
         self.I_max = 255
-        self.N_leds = 64
         self.setEnabled = False
-        self.intesnsity=0
+        self.intensity=0
 
-        self.pattern = np.array((np.reshape(np.random.randint(0,self.I_max ,self.N_leds**2),(self.N_leds,self.N_leds)),
-                       np.reshape(np.random.randint(0,self.I_max ,self.N_leds**2),(self.N_leds,self.N_leds)),
-                       np.reshape(np.random.randint(0,self.I_max ,self.N_leds**2),(self.N_leds,self.N_leds))))
-        
+        try:
+            self.Nx = LEDMatrixInfo.managerProperties['Nx']
+            self.Ny = LEDMatrixInfo.managerProperties['Ny']
+        except:
+            self.Nx = 8
+            self.Ny = 8
+
+        self.NLeds = self.Nx*self.Ny
 
         self._rs232manager = lowLevelManagers['rs232sManager'][
             LEDMatrixInfo.managerProperties['rs232device']
         ]
-            
-        self.esp32 = self._rs232manager._esp32
+       
+        # initialize the LEDMatrix device that holds all necessary states^
+        self.mLEDmatrix = self._rs232manager._esp32.led
+        self.mLEDmatrix.setLEDArrayConfig(ledArrPin=None, ledArrNum=self.NLeds)
+
         super().__init__(LEDMatrixInfo, name, isBinary=False, valueUnits='mW', valueDecimals=0)
 
-    def setAll(self, intensity=(0,0,0)):
-        self.intesnsity=intensity
-        self.esp32.send_LEDMatrix_full(intensity=intensity,timeout=1)
-    
+    def setAll(self, state=(0,0,0), intensity=None):
+        # dealing with on or off,
+        # intensity is adjjusting the global value
+        self.mLEDmatrix.setAll(state, intensity)
+
     def setPattern(self, pattern):
-        self.pattern=np.int16(pattern).T
-        # assuming flat array
-        #if len(self.pattern)!=3:
-        #    self.pattern=np.reshape(np.transpose(self.pattern), (3,int(np.sqrt(self.N_leds)),int(np.sqrt(self.N_leds))))
-        self.esp32.send_LEDMatrix_array(self.pattern)
-        
+        self.mLEDmatrix.pattern(pattern)
+    
+    def getPattern(self):
+        return self.mLEDmatrix.getPattern()
+
     def setEnabled(self, enabled):
         """Turn on (N) or off (F) LEDMatrix emission"""
         self.setEnabled = enabled
-        #self.esp32.setLEDMatrixPattern(self.pattern*self.setEnabled)
-
-    def setLEDSingle(self, indexled=0, intensity=(255,255,255)):
+        
+    def setLEDSingle(self, indexled=0, state=(0,0,0)):
         """Handles output power.
         Sends a RS232 command to the LEDMatrix specifying the new intensity.
         """
-        self.esp32.send_LEDMatrix_single(indexled, intensity, timeout=1)
-        
+        self.mLEDmatrix.setSingle(indexled, state=state)
+
+    def setLEDIntensity(self, intensity=(0,0,0)):
+        self.mLEDmatrix.setIntensity(intensity)
+
+
 
 # Copyright (C) 2020-2021 ImSwitch developers
 # This file is part of ImSwitch.
