@@ -2,14 +2,138 @@ import enum
 import os
 import time
 from io import BytesIO
+from typing import Dict, Optional, Type
 
 import h5py
+import zarr
 import numpy as np
 import tifffile as tiff
-import cv2
 
 from imswitch.imcommon.framework import Signal, SignalInterface, Thread, Worker
 from imswitch.imcommon.model import initLogger
+import abc
+import logging
+
+from imswitch.imcontrol.model.managers.DetectorsManager import DetectorsManager
+
+logger = logging.getLogger(__name__)
+
+
+class AsTemporayFile(object):
+    """ A temporary file that when exiting the context manager is renamed to its original name. """
+
+    def __init__(self, filepath, tmp_extension='.tmp'):
+        if os.path.exists(filepath):
+            raise FileExistsError(f'File {filepath} already exists.')
+        self.path = filepath
+        self.tmp_path = filepath + tmp_extension
+
+    def __enter__(self):
+        return self.tmp_path
+
+    def __exit__(self, *args, **kwargs):
+        os.rename(self.tmp_path, self.path)
+        logger.info("Renamed file from %s to %s", self.tmp_path, self.path)
+
+
+
+class Storer(abc.ABC):
+    """ Base class for storing data"""
+
+    def __init__(self, filepath, detectorManager):
+        self.filepath = filepath
+        self.detectorManager: DetectorsManager = detectorManager
+
+    def snap(self, images: Dict[str, np.ndarray], attrs: Dict[str, str] = None):
+        """ Stores images and attributes according to the spec of the storer """
+        raise NotImplementedError
+
+    def stream(self, data = None, **kwargs):
+        """ Stores data in a streaming fashion. """
+        raise NotImplementedError
+
+
+
+
+class ZarrStorer(Storer):
+    """ A storer that stores the images in a zarr file store """
+
+    def snap(self, images: Dict[str, np.ndarray], attrs: Dict[str, str] = None):
+
+        with AsTemporayFile(f'{self.filepath}.zarr') as path:
+            store = zarr.storage.DirectoryStore(path)
+            root = zarr.group(store=store)
+
+            for channel, image in images.items():
+                shape = self.detectorManager[channel].shape
+
+                d = root.create_dataset(channel, data=image, shape=tuple(reversed(shape)),
+                                        chunks=(512, 512), dtype='i2') #TODO: why not dynamic chunking?
+                d.attrs["ImSwitchData"] = attrs[channel]
+                logger.info(f"Saved image to zarr file {path}")
+
+
+class HDF5Storer(Storer):
+    """ A storer that stores the images in a series of hd5 files """
+
+    def snap(self, images: Dict[str, np.ndarray], attrs: Dict[str, str] = None):
+
+        for channel, image in images.items():
+
+            with AsTemporayFile(f'{self.filepath}_{channel}.h5') as path:
+                file = h5py.File(path, 'w')
+
+                shape = self.detectorManager[channel].shape
+
+                dataset = file.create_dataset('data', tuple(reversed(shape)), dtype='i2')
+
+                for key, value in attrs[channel].items():
+                    try:
+                        dataset.attrs[key] = value
+                    except:
+                        logger.debug(f'Could not put key:value pair {key}:{value} in hdf5 metadata.')
+
+                dataset.attrs['detector_name'] = channel
+
+                # For ImageJ compatibility
+                dataset.attrs['element_size_um'] = \
+                    self.detectorManager[channel].pixelSizeUm
+
+                dataset[:, ...] = np.moveaxis(image, 0, -1)
+
+                file.close()
+
+
+class TiffStorer(Storer):
+    """ A storer that stores the images in a series of tiff files """
+
+    def snap(self, images: Dict[str, np.ndarray], attrs: Dict[str, str] = None):
+        for channel, image in images.items():
+            with AsTemporayFile(f'{self.filepath}_{channel}.tiff') as path:
+                tiff.imwrite(path, image,) # TODO: Parse metadata to tiff meta data
+
+
+
+class SaveMode(enum.Enum):
+    Disk = 1
+    RAM = 2
+    DiskAndRAM = 3
+    Numpy = 4
+
+
+class SaveFormat(enum.Enum):
+    HDF5 = 1
+    TIFF = 2
+    ZARR = 3
+
+
+DEFAULT_STORER_MAP: Dict[str, Type[Storer]] = {
+    SaveFormat.ZARR: ZarrStorer,
+    SaveFormat.HDF5: HDF5Storer,
+    SaveFormat.TIFF: TiffStorer
+}
+
+
 
 
 class RecordingManager(SignalInterface):
@@ -27,10 +151,10 @@ class RecordingManager(SignalInterface):
         str, object, object, bool
     )  # (name, file, filePath, savedToDisk)
 
-    def __init__(self, detectorsManager):
+    def __init__(self, detectorsManager, storerMap: Optional[Dict[str, Type[Storer]]] = None):
         super().__init__()
         self.__logger = initLogger(self)
-
+        self.__storerMap = storerMap or DEFAULT_STORER_MAP
         self._memRecordings = {}  # { filePath: bytesIO }
         self.__detectorsManager = detectorsManager
         self.__record = False
@@ -53,8 +177,8 @@ class RecordingManager(SignalInterface):
     def detectorsManager(self):
         return self.__detectorsManager
 
-    def startRecording(self, detectorNames, recMode, savename, saveMode, saveFormat, attrs,
-                       singleMultiDetectorFile=False, singleLapseFile=False,
+    def startRecording(self, detectorNames, recMode, savename, saveMode, attrs,
+                       saveFormat=SaveFormat.HDF5, singleMultiDetectorFile=False, singleLapseFile=False,
                        recFrames=None, recTime=None):
         """ Starts a recording with the specified detectors, recording mode,
         file name prefix and attributes to save to the recording per detector.
@@ -84,7 +208,7 @@ class RecordingManager(SignalInterface):
         method will wait until the recording is complete before returning. """
 
         self.__detectorsManager.execOnAll(lambda c: c.flushBuffers(),
-                                    condition=lambda c: c.forAcquisition)
+                                          condition=lambda c: c.forAcquisition)
 
         if self.__record:
             self.__logger.info('Stopping recording')
@@ -95,68 +219,38 @@ class RecordingManager(SignalInterface):
         if wait:
             self.__thread.wait()
 
-
     def snap(self, detectorNames, savename, saveMode, saveFormat, attrs):
         """ Saves an image with the specified detectors to a file
         with the specified name prefix, save mode, file format and attributes
         to save to the capture per detector. """
         acqHandle = self.__detectorsManager.startAcquisition()
+
+
         try:
             images = {}
+
+            # Acquire data
             for detectorName in detectorNames:
                 images[detectorName] = self.__detectorsManager[detectorName].getLatestFrame(is_save=True)
 
-            for detectorName in detectorNames:
-                image = images[detectorName]
+            if saveFormat:
+                storer = self.__storerMap[saveFormat]
 
-                if saveMode == SaveMode.Numpy:
-                    return
+                if saveMode == SaveMode.Disk or saveMode == SaveMode.DiskAndRAM:
+                    # Save images to disk
+                    store = storer(savename, self.__detectorsManager)
+                    store.snap(images, attrs)
 
-                fileExtension = str(saveFormat.name).lower()
-                filePath = self.getSaveFilePath(f'{savename}_{detectorName}.{fileExtension}')
 
-                if saveMode != SaveMode.RAM:
-                    # Write file
-                    if saveFormat == SaveFormat.HDF5:
-                        file = h5py.File(filePath, 'w')
-
-                        shape = self.__detectorsManager[detectorName].shape
-                        dataset = file.create_dataset('data', tuple(reversed(shape)), dtype='i2')
-
-                        for key, value in attrs[detectorName].items():
-                            #self.__logger.debug(key)
-                            #self.__logger.debug(value)
-                            try:
-                                dataset.attrs[key] = value
-                            except:
-                                self.__logger.debug(f'Could not put key:value pair {key}:{value} in hdf5 metadata.')
-
-                        dataset.attrs['detector_name'] = detectorName
-
-                        # For ImageJ compatibility
-                        dataset.attrs['element_size_um'] =\
-                            self.__detectorsManager[detectorName].pixelSizeUm
-                        
-                        dataset[:,...] = np.moveaxis(image,0,-1)
-                        file.close()
-                    elif saveFormat == SaveFormat.TIFF or self.saveFormat == SaveFormat.TIFF_Single:
-                        tiff.imwrite(filePath, image)
-                    else:
-                        raise ValueError(f'Unsupported save format "{saveFormat}"')
-
-                # Handle memory snaps
                 if saveMode == SaveMode.RAM or saveMode == SaveMode.DiskAndRAM:
-                    name = os.path.basename(f'{savename}_{detectorName}')
-                    self.sigMemorySnapAvailable.emit(name, image, filePath,
-                                                     saveMode == SaveMode.DiskAndRAM)
-
-
+                    for channel, image in images.items():
+                        name = os.path.basename(f'{savename}_{channel}')
+                        self.sigMemorySnapAvailable.emit(name, image, savename, saveMode == SaveMode.DiskAndRAM)
 
         finally:
             self.__detectorsManager.stopAcquisition(acqHandle)
             if saveMode == SaveMode.Numpy:
-                return image
-
+                return images
 
     def snapImagePrev(self, detectorName, savename, saveFormat, image, attrs):
         """ Saves a previously taken image to a file with the specified name prefix,
@@ -172,8 +266,6 @@ class RecordingManager(SignalInterface):
             dataset = file.create_dataset('data', tuple(reversed(shape)), dtype='i2')
 
             for key, value in attrs[detectorName].items():
-                #self.__logger.debug(key)
-                #self.__logger.debug(value)
                 try:
                     dataset.attrs[key] = value
                 except:
@@ -182,17 +274,24 @@ class RecordingManager(SignalInterface):
             dataset.attrs['detector_name'] = detectorName
 
             # For ImageJ compatibility
-            dataset.attrs['element_size_um'] =\
+            dataset.attrs['element_size_um'] = \
                 self.__detectorsManager[detectorName].pixelSizeUm
-            
-            #dataset[:,:,:] = image
-            dataset[:,...] = np.moveaxis(image,0,-1)
+
+            dataset[:, ...] = np.moveaxis(image, 0, -1)
             file.close()
         elif saveFormat == SaveFormat.TIFF:
             tiff.imwrite(filePath, image)
+        if saveFormat == SaveFormat.ZARR:
+            path = self.getSaveFilePath(f'{savename}.{fileExtension}')
+            store = zarr.storage.DirectoryStore(path)
+            root = zarr.group(store=store)
+            shape = self.__detectorsManager[detectorName].shape
+            d = root.create_dataset(detectorName, data=image, shape=tuple(reversed(shape)), chunks=(512, 512),
+                                    dtype='i2')
+            d.attrs["ImSwitchData"] = attrs[detectorName]
+            store.close()
         else:
-            raise ValueError(f'Unsupported save format "{saveFormat}"')    
-
+            raise ValueError(f'Unsupported save format "{saveFormat}"')
 
     def getSaveFilePath(self, path, allowOverwriteDisk=False, allowOverwriteMem=False):
         newPath = path
@@ -228,7 +327,7 @@ class RecordingWorker(Worker):
             self.__recordingManager.detectorsManager.stopAcquisition(acqHandle)
 
     def _record(self):
-        if self.saveFormat == SaveFormat.HDF5:
+        if self.saveFormat == SaveFormat.HDF5 or self.saveFormat == SaveFormat.ZARR:
             files, fileDests, filePaths = self._getFiles()
 
         shapes = {detectorName: self.__recordingManager.detectorsManager[detectorName].shape
@@ -266,7 +365,12 @@ class RecordingWorker(Worker):
                 )
 
                 for key, value in self.attrs[detectorName].items():
-                    datasets[detectorName].attrs[key] = value
+                    self.__logger.debug(key)
+                    self.__logger.debug(value)
+                    try:
+                        datasets[detectorName].attrs[key] = value
+                    except:
+                        pass
 
                 datasets[detectorName].attrs['detector_name'] = detectorName
 
@@ -275,7 +379,10 @@ class RecordingWorker(Worker):
                     = self.__recordingManager.detectorsManager[detectorName].pixelSizeUm
 
                 for key, value in self.attrs[detectorName].items():
-                    datasets[detectorName].attrs[key] = value
+                    try:
+                        datasets[detectorName].attrs[key] = value
+                    except:
+                        pass
 
             elif self.saveFormat == SaveFormat.MP4:
                 # Need to initiliaze videowriter for each detector
@@ -289,10 +396,21 @@ class RecordingWorker(Worker):
                 self.__logger.debug(filePath)
 
             elif self.saveFormat == SaveFormat.TIFF:
-                # Need to initiliaze TIF writer?
                 fileExtension = str(self.saveFormat.name).lower()
                 filenames[detectorName] = self.__recordingManager.getSaveFilePath(
                     f'{self.savename}_{detectorName}.{fileExtension}', False, False)
+
+            elif self.saveFormat == SaveFormat.ZARR:
+                datasets[detectorName] = files[detectorName].create_dataset(datasetName, shape=(1, *reversed(shape)),
+                                                                            dtype='i2', chunks=(1, 512, 512)
+                                                                            )
+
+                datasets[detectorName].attrs['ImSwitchData'] = self.attrs[detectorName]
+                datasets[detectorName].attrs['detector_name'] = detectorName
+                # For ImageJ compatibility
+                datasets[detectorName].attrs['element_size_um'] \
+                    = self.__recordingManager.detectorsManager[detectorName].pixelSizeUm
+                datasets[detectorName].attrs['writing'] = True
 
         self.__recordingManager.sigRecordingStarted.emit()
         try:
@@ -317,9 +435,9 @@ class RecordingWorker(Worker):
 
                         if n > 0:
                             it = currentFrame[detectorName]
-                            if self.saveFormat == SaveFormat.TIFF or self.saveFormat == SaveFormat.TIFF_Single:
+                            if self.saveFormat == SaveFormat.TIFF:
                                 try:
-                                    filePath = filenames[detectorName] 
+                                    filePath = filenames[detectorName]
                                     tiff.imwrite(filePath, newFrames, append=True)
                                 except ValueError:
                                     self.__logger.error("TIFF File exceeded 4GB.")
@@ -337,6 +455,15 @@ class RecordingWorker(Worker):
                                     dataset.resize(recFrames, axis=0)
                                     dataset[it:recFrames, :, :] = newFrames[0:recFrames - it]
                                     currentFrame[detectorName] = recFrames
+                            elif self.saveFormat == SaveFormat.ZARR:
+                                dataset = datasets[detectorName]
+                                if it == 0:
+                                    dataset[0, :, :] = newFrames[0, :, :]
+                                    if n > 0:
+                                        dataset.append(newFrames[1:n, :, :])
+                                else:
+                                    dataset.append(newFrames)
+                                currentFrame[detectorName] += n
 
                             # Things get a bit weird if we have multiple detectors when we report
                             # the current frame number, since the detectors may not be synchronized.
@@ -360,17 +487,17 @@ class RecordingWorker(Worker):
                         newFrames = self._getNewFrames(detectorName)
                         n = len(newFrames)
                         if n > 0:
-                            if self.saveFormat == SaveFormat.TIFF or self.saveFormat == SaveFormat.TIFF_Single:
+                            if self.saveFormat == SaveFormat.TIFF:
                                 try:
-                                    filePath = filenames[detectorName] 
-                                    tiff.imwrite(filePath, newFrames, append=True) # TODO: Single Tiff export doesnt work! 
+                                    filePath = filenames[detectorName]
+                                    tiff.imwrite(filePath, newFrames, append=True)
                                 except ValueError:
                                     self.__logger.error("TIFF File exceeded 4GB.")
                                     if self.saveFormat == SaveFormat.TIFF:
                                         filePath = self.__recordingManager.getSaveFilePath(
                                             f'{self.savename}_{detectorName}.{fileExtension}', False, False)
                                         continue
-                            elif self.saveFormat == SaveFormat.HDF5:
+                            elif self.saveFormat == SaveFormat.HDF5 or self.saveFormat == SaveFormat.ZARR:
                                 it = currentFrame[detectorName]
                                 dataset = datasets[detectorName]
                                 dataset.resize(n + it, axis=0)
@@ -397,9 +524,9 @@ class RecordingWorker(Worker):
                         newFrames = self._getNewFrames(detectorName)
                         n = len(newFrames)
                         if n > 0:
-                            if self.saveFormat == SaveFormat.TIFF or self.saveFormat == SaveFormat.TIFF_Single:
+                            if self.saveFormat == SaveFormat.TIFF:
                                 try:
-                                    filePath = filenames[detectorName] 
+                                    filePath = filenames[detectorName]
                                     tiff.imwrite(filePath, newFrames, append=True)
                                 except ValueError:
                                     self.__logger.error("TIFF File exceeded 4GB.")
@@ -413,17 +540,16 @@ class RecordingWorker(Worker):
                                 dataset = datasets[detectorName]
                                 dataset.resize(n + it, axis=0)
                                 dataset[it:it + n, :, :] = newFrames
-                            elif self.saveFormat == SaveFormat.MP4:
-                                for iframe in range(n):
-                                    frame = newFrames[iframe,:,:]
-                                    self.__logger.debug(frame.shape)
-                                    self.__logger.debug(type(frame))
-                                    self.__logger.debug(datasets[detectorName])
-                                    #https://stackoverflow.com/questions/30509573/writing-an-mp4-video-using-python-opencv
-                                    frame = cv2.cvtColor(cv2.convertScaleAbs(frame), cv2.COLOR_GRAY2BGR)
-                                    self.__logger.debug(type(frame))
 
-                                    datasets[detectorName].write(frame)
+                            elif self.saveFormat == SaveFormat.ZARR:
+                                it = currentFrame[detectorName]
+                                dataset = datasets[detectorName]
+                                if it == 0:
+                                    dataset[0, :, :] = newFrames[0, :, :]
+                                    if n > 0:
+                                        dataset.append(newFrames[1:n, :, :])
+                                else:
+                                    dataset.append(newFrames)
 
                             currentFrame[detectorName] += n
 
@@ -438,15 +564,12 @@ class RecordingWorker(Worker):
                 raise ValueError('Unsupported recording mode specified')
         finally:
 
-            if self.saveFormat == SaveFormat.MP4:
-                for detectorName, file in files.items():
-                    datasets[detectorName].release()
-
-            if self.saveFormat == SaveFormat.HDF5:
+            if self.saveFormat == SaveFormat.HDF5 or self.saveFormat == SaveFormat.ZARR:
                 for detectorName, file in files.items():
                     # Remove default frame if no frames have been captured
                     if currentFrame[detectorName] < 1:
-                        datasets[detectorName].resize(0, axis=0)
+                        if self.saveFormat == SaveFormat.HDF5:
+                            datasets[detectorName].resize(0, axis=0)
 
                     # Handle memory recordings
                     if self.saveMode == SaveMode.RAM or self.saveMode == SaveMode.DiskAndRAM:
@@ -463,7 +586,11 @@ class RecordingWorker(Worker):
                                 name, file, filePath, True
                             )
                     else:
-                        file.close()
+                        if self.saveFormat == SaveFormat.HDF5:
+                            file.close()
+                        else:
+                            datasets[detectorName].attrs['writing'] = False
+                            self.store.close()
 
             self.__recordingManager.endRecording(wait=False)
 
@@ -474,11 +601,13 @@ class RecordingWorker(Worker):
         files = {}
         fileDests = {}
         filePaths = {}
+        extension = 'hdf5' if self.saveFormat == SaveFormat.HDF5 else 'zarr'
+
         for detectorName in self.detectorNames:
             if singleMultiDetectorFile:
-                baseFilePath = f'{self.savename}.hdf5'
+                baseFilePath = f'{self.savename}.{extension}'
             else:
-                baseFilePath = f'{self.savename}_{detectorName}.hdf5'
+                baseFilePath = f'{self.savename}_{detectorName}.{extension}'
 
             filePaths[detectorName] = self.__recordingManager.getSaveFilePath(
                 baseFilePath,
@@ -499,8 +628,12 @@ class RecordingWorker(Worker):
             if singleMultiDetectorFile and len(files) > 0:
                 files[detectorName] = list(files.values())[0]
             else:
-                files[detectorName] = h5py.File(fileDests[detectorName],
-                                                'a' if singleLapseFile else 'w-')
+                if self.saveFormat == SaveFormat.HDF5:
+                    files[detectorName] = h5py.File(fileDests[detectorName],
+                                                    'a' if singleLapseFile else 'w-')
+                elif self.saveFormat == SaveFormat.ZARR:
+                    self.store = zarr.storage.DirectoryStore(fileDests[detectorName])
+                    files[detectorName] = zarr.group(store=self.store, overwrite=True)
 
         return files, fileDests, filePaths
 
@@ -518,18 +651,6 @@ class RecMode(enum.Enum):
     UntilStop = 5
 
 
-class SaveMode(enum.Enum):
-    Disk = 1
-    RAM = 2
-    DiskAndRAM = 3
-    Numpy = 4
-
-
-class SaveFormat(enum.Enum):
-    HDF5 = 1
-    TIFF = 2
-    TIFF_Single = 3
-    MP4 = 4
 
 
 # Copyright (C) 2020-2021 ImSwitch developers
