@@ -1,13 +1,17 @@
+from .AutofocusManager import AutofocusManager
 from .DetectorsManager import DetectorsManager, NoDetectorsError
 from .LasersManager import LasersManager
 from .LEDMatrixsManager import LEDMatrixsManager
 from .MultiManager import MultiManager
-from .NidaqManager import NidaqManager
-#from .PulseStreamerManager import PulseStreamerManager
 from .PositionersManager import PositionersManager
 from .RS232sManager import RS232sManager
+from .OFMsManager import OFMsManager
 from .RecordingManager import RecordingManager, RecMode, SaveMode, SaveFormat
 from .SLMManager import SLMManager
+from .UC2ConfigManager import UC2ConfigManager
 from .SIMManager import SIMManager
 from .MCTManager import MCTManager
+from .HistoScanManager import HistoScanManager
+from .PixelCalibrationManager import PixelCalibrationManager
+from .ISMManager import ISMManager
 from .ScanManager import ScanManager
