@@ -1,6 +1,7 @@
 from imswitch.imcommon.model import initLogger
 from .PositionerManager import PositionerManager
-
+import time
+import numpy as np
 
 PHYS_FACTOR = 1
 class ESP32StageManager(PositionerManager):
@@ -15,56 +16,193 @@ class ESP32StageManager(PositionerManager):
         ]
         self.__logger = initLogger(self, instanceName=name)
 
-        self.is_enabled = False
-        self.backlash_x = 0
-        self.backlash_y = 0
-        self.backlash_z= 0 # TODO: Map that to the JSON!
+        # calibrated stepsize in steps/µm
+        if positionerInfo.managerProperties.get('stepsizeX') is not None:
+            self.stepsizeX = positionerInfo.managerProperties['stepsizeX']
+        else:
+            self.stepsizeX = 1
+            
+        # calibrated stepsize
+        if positionerInfo.managerProperties.get('stepsizeY') is not None:
+            self.stepsizeY = positionerInfo.managerProperties['stepsizeY']
+        else:
+            self.stepsizeY = 1
 
-    def move(self, value=0, axis="X", is_absolute=False, is_blocking=True):
+        # calibrated stepsize
+        if positionerInfo.managerProperties.get('stepsizeZ') is not None:
+            self.stepsizeZ = positionerInfo.managerProperties['stepsizeZ']
+        else:
+            self.stepsizeZ = 1
+            
+        # calibrated stepsize
+        if positionerInfo.managerProperties.get('stepsizeT') is not None:
+            self.stepsizeT = positionerInfo.managerProperties['stepsizeT']
+        else:
+            self.stepsizeT = 1            
+
+        # miniumum/maximum steps in X
+        if positionerInfo.managerProperties.get('minX') is not None:
+            self.minX = positionerInfo.managerProperties['minX']
+        else:
+            self.minX = -np.inf
+        if positionerInfo.managerProperties.get('maxX') is not None:
+            self.maxX = positionerInfo.managerProperties['maxX']
+        else:
+            self.maxX = np.inf
+            
+        # minimum/maximum steps in Y
+        if positionerInfo.managerProperties.get('minY') is not None:
+            self.minY = positionerInfo.managerProperties['minY']
+        else:
+           self.minY = -np.inf
+        if positionerInfo.managerProperties.get('maxY') is not None:
+            self.maxY = positionerInfo.managerProperties['maxY']
+        else:
+           self.maxY = np.inf
+        # minimum/maximum steps in Z
+        if positionerInfo.managerProperties.get('minZ') is not None:
+            self.minZ = positionerInfo.managerProperties['minZ']
+        else:
+           self.minZ = -np.inf
+        if positionerInfo.managerProperties.get('maxZ') is not None:
+            self.maxZ = positionerInfo.managerProperties['maxZ']
+        else:
+           self.maxZ = np.inf           
+        # minimum/maximum steps in T
+        if positionerInfo.managerProperties.get('minT') is not None:
+            self.minT = positionerInfo.managerProperties['minT']
+        else:
+           self.minT = -np.inf
+        if positionerInfo.managerProperties.get('maxT') is not None:
+            self.maxT = positionerInfo.managerProperties['maxT']
+        else:
+           self.maxT = np.inf       
+           
+        # calibrated backlash
+        if positionerInfo.managerProperties.get('backlashX') is not None:
+            self.backlashX = positionerInfo.managerProperties['backlashX']
+        else:
+            self.backlashX = 1
+
+        # calibrated backlash
+        if positionerInfo.managerProperties.get('backlashY') is not None:
+            self.backlashY = positionerInfo.managerProperties['backlashY']
+        else:
+            self.backlashY = 1
+
+        # calibrated backlash
+        if positionerInfo.managerProperties.get ('backlashZ') is not None:
+            self.backlashZ = positionerInfo.managerProperties['backlashZ']
+        else:
+            self.backlashZ = 1
+
+        # calibrated backlash
+        if positionerInfo.managerProperties.get ('backlashT') is not None:
+            self.backlashT = positionerInfo.managerProperties['backlashT']
+        else:
+            self.backlashT = 1
+            
+        # grab motor object
+        self._motor = self._rs232manager._esp32.motor
+            
+        # setup motors
+        self.setupMotor(self.minX, self.maxX, self.stepsizeX, self.backlashX, "X")
+        self.setupMotor(self.minY, self.maxY, self.stepsizeY, self.backlashY, "Y")
+        self.setupMotor(self.minZ, self.maxZ, self.stepsizeZ, self.backlashZ, "Z")
+        self.setupMotor(self.minT, self.maxT, self.stepsizeT, self.backlashT, "T")
+        
+        self.is_enabled = False
+
+
+    def setupMotor(self, minPos, maxPos, stepSize, backlash, axis):
+        self._motor.setup_motor(axis=axis, minPos=minPos, maxPos=maxPos, stepSize=stepSize, backlash=backlash)
+        
+    def move(self, value=0, axis="X", speed=None, is_absolute=False, is_blocking=False, timeout=np.inf):
+        if speed is None:
+            if axis == "X": speed = self.speed["X"]
+            if axis == "Y": speed = self.speed["Y"]
+            if axis == "Z": speed = self.speed["Z"]
+            if axis == "XY": speed = (self.speed["X"], self.speed["Y"])
+            if axis == "XYZ": speed = (self.speed["X"], self.speed["Y"], self.speed["Z"])
+            
         if axis == 'X':
-            self._rs232manager._esp32.move_x(value, self.speed["X"], is_absolute=is_absolute, is_enabled=self.is_enabled, is_blocking=is_blocking)
-            self._position[axis] = self._position[axis] + value
+            self._motor.move_x(value, speed, is_absolute=is_absolute, is_enabled=self.is_enabled, is_blocking=is_blocking, timeout=timeout)
+            if not is_absolute: self._position[axis] = self._position[axis] + value
+            else: self._position[axis] = value
         elif axis == 'Y':
-            self._rs232manager._esp32.move_y(value, self.speed["Y"], is_absolute=is_absolute, is_enabled=self.is_enabled, is_blocking=is_blocking)
-            self._position[axis] = self._position[axis] + value
+            self._motor.move_y(value, speed, is_absolute=is_absolute, is_enabled=self.is_enabled, is_blocking=is_blocking, timeout=timeout)
+            if not is_absolute: self._position[axis] = self._position[axis] + value
+            else: self._position[axis] = value
         elif axis == 'Z':
-            self._rs232manager._esp32.move_z(value, self.speed["Z"], is_absolute=is_absolute, is_enabled=self.is_enabled, is_blocking=is_blocking)
-            self._position[axis] = self._position[axis] + value
+            self._motor.move_z(value, speed, is_absolute=is_absolute, is_enabled=self.is_enabled, is_blocking=is_blocking, timeout=timeout)
+            if not is_absolute: self._position[axis] = self._position[axis] + value
+            else: self._position[axis] = value
+        elif axis == 'XY':
+            self._motor.move_xy(value, speed, is_absolute=is_absolute, is_enabled=self.is_enabled, is_blocking=is_blocking, timeout=timeout)
+            for i, iaxis in enumerate(("X", "Y")):
+                if not is_absolute: self._position[iaxis] = self._position[iaxis] + value[i]
+                else: self._position[iaxis] = value[i]
         elif axis == 'XYZ':
-            self._rs232manager._esp32.move_xyz(value, self.speed, is_absolute=is_absolute, is_enabled=self.is_enabled, is_blocking=is_blocking)
-            self._position["X"] = self._position["X"] + value[0]
-            self._position["Y"] = self._position["Y"] + value[1]
-            self._position["Z"] = self._position["Z"] + value[2]
+            self._motor.move_xyz(value, speed, is_absolute=is_absolute, is_enabled=self.is_enabled, is_blocking=is_blocking, timeout=timeout)
+            for i, iaxis in enumerate(("X", "Y")):
+                if not is_absolute: self._position[iaxis] = self._position[iaxis] + value[i]
+                else: self._position[iaxis] = value[i]
+
         else:
             print('Wrong axis, has to be "X" "Y" or "Z".')
             return
+
     
     def measure(self, sensorID=0, NAvg=100):
-        return self._rs232manager._esp32.read_sensor(sensorID=sensorID, NAvg=NAvg)
+        return self._motor.read_sensor(sensorID=sensorID, NAvg=NAvg)
 
     def setupPIDcontroller(self, PIDactive=1, Kp=100, Ki=10, Kd=1, target=500, PID_updaterate=200):
-        return self._rs232manager._esp32.set_pidcontroller(PIDactive=PIDactive, Kp=Kp, Ki=Ki, Kd=Kd, target=target, PID_updaterate=PID_updaterate)
+        return self._motor.set_pidcontroller(PIDactive=PIDactive, Kp=Kp, Ki=Ki, Kd=Kd, target=target, PID_updaterate=PID_updaterate)
 
     def moveForever(self, speed=(0,0,0), is_stop=False):
-        self._rs232manager._esp32.move_forever(speed=speed, is_stop=is_stop)
+        self._motor.move_forever(speed=speed, is_stop=is_stop)
         
     def setEnabled(self, is_enabled):
         self.is_enabled = is_enabled
 
-    def setSpeed(self, speed):
-        self._speed = speed
+    def setSpeed(self, speed, axis=None):
+        #TODO: Map that to the JSON!
+        if type(speed)==int and axis == None:
+            self._speed["X"] = speed
+            self._speed["Y"] = speed
+            self._speed["Z"] = speed
+        else:
+            self._speed[axis] = speed
 
     def setPosition(self, value, axis):
         if value: value+=1 # TODO: Firmware weirdness
-        self._rs232manager._esp32.set_position(axis=axis, position=value)
+        self._motor.set_position(axis=axis, position=value)
         self._position[axis] = value
 
     def closeEvent(self):
         pass
 
     def get_abs(self, axis=1):
-        abspos = self._rs232manager._esp32.get_position(axis=axis)
+        abspos = self._motor.get_position(axis=axis)
         return abspos
+    
+    def home_x(self):
+        self._motor.home_x()
+        self._position["X"] = 0
+        
+    def home_y(self):
+        self._motor.home_y()
+        self._position["Y"] = 0
+        
+    def home_z(self):
+        self._motor.home_z()
+        self._position["Z"] = 0
+        
+    def home_xyz(self):
+        self._motor.home_xyz()
+        self._position["X"] = 0
+        self._position["Y"] = 0
+        self._position["Z"] = 0
 
 
 
