@@ -9,7 +9,7 @@ class PositionerWidget(Widget):
 
     sigStepUpClicked = QtCore.Signal(str, str)  # (positionerName, axis)
     sigStepDownClicked = QtCore.Signal(str, str)  # (positionerName, axis)
-    sigsetSpeedClicked = QtCore.Signal(str)  # (speed, axis)
+    sigsetSpeedClicked = QtCore.Signal(str, str)  # (positionerName, axis)
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -18,7 +18,7 @@ class PositionerWidget(Widget):
         self.grid = QtWidgets.QGridLayout()
         self.setLayout(self.grid)
 
-    def addPositioner(self, positionerName, axes):
+    def addPositioner(self, positionerName, axes, hasSpeed):
         for i in range(len(axes)):
             axis = axes[i]
             parNameSuffix = self._getParNameSuffix(positionerName, axis)
@@ -40,25 +40,6 @@ class PositionerWidget(Widget):
             self.grid.addWidget(QtWidgets.QLabel('Step'), self.numPositioners, 4)
             self.grid.addWidget(self.pars['StepEdit' + parNameSuffix], self.numPositioners, 5)
             self.grid.addWidget(self.pars['StepUnit' + parNameSuffix], self.numPositioners, 6)
-            
-            self.pars['Speed' + parNameSuffix] = QtWidgets.QLabel(f'<strong>{0:.2f} µm</strong>')
-            self.pars['Speed' + parNameSuffix].setTextFormat(QtCore.Qt.RichText)
-            self.pars['ButtonSpeedEnter' + parNameSuffix] = guitools.BetterPushButton('Enter')
-            self.pars['SpeedEdit' + parNameSuffix] = QtWidgets.QLineEdit('1000')
-            self.pars['SpeedUnit' + parNameSuffix] = QtWidgets.QLabel(' µm/s')
-
-            self.grid.addWidget(QtWidgets.QLabel('Speed' + parNameSuffix), self.numPositioners, 9)
-            self.grid.addWidget(self.pars['SpeedEdit' + parNameSuffix], self.numPositioners, 10)
-            self.grid.addWidget(self.pars['SpeedUnit' + parNameSuffix], self.numPositioners, 11)
-            self.grid.addWidget(self.pars['ButtonSpeedEnter' + parNameSuffix], self.numPositioners, 12)
-            self.grid.addWidget(self.pars['Speed' + parNameSuffix], self.numPositioners, 7)
-                
-            
-            self.pars['ButtonSpeedEnter' + parNameSuffix].clicked.connect(
-                lambda *args: self.sigsetSpeedClicked.emit(positionerName, axis)
-            )
-
-            self.numPositioners += 1
 
             # Connect signals
             self.pars['UpButton' + parNameSuffix].clicked.connect(
@@ -68,6 +49,23 @@ class PositionerWidget(Widget):
                 lambda *args, axis=axis: self.sigStepDownClicked.emit(positionerName, axis)
             )
 
+            if hasSpeed:
+                self.pars['Speed' + parNameSuffix] = QtWidgets.QLabel(f'<strong>{0:.2f} µm/s</strong>')
+                self.pars['Speed' + parNameSuffix].setTextFormat(QtCore.Qt.RichText)
+                self.pars['ButtonSpeedEnter' + parNameSuffix] = guitools.BetterPushButton('Enter')
+                self.pars['SpeedEdit' + parNameSuffix] = QtWidgets.QLineEdit('1000')
+                self.pars['SpeedUnit' + parNameSuffix] = QtWidgets.QLabel(' µm/s')
+                self.grid.addWidget(self.pars['SpeedEdit' + parNameSuffix], self.numPositioners, 10)
+                self.grid.addWidget(self.pars['SpeedUnit' + parNameSuffix], self.numPositioners, 11)
+                self.grid.addWidget(self.pars['ButtonSpeedEnter' + parNameSuffix], self.numPositioners, 12)
+                self.grid.addWidget(self.pars['Speed' + parNameSuffix], self.numPositioners, 7)
+
+
+                self.pars['ButtonSpeedEnter'+ parNameSuffix].clicked.connect(
+                    lambda *args, axis=axis: self.sigsetSpeedClicked.emit(positionerName, axis)
+                )
+
+            self.numPositioners += 1
 
     def getStepSize(self, positionerName, axis):
         """ Returns the step size of the specified positioner axis in
