@@ -7,20 +7,26 @@ from .EtSTEDWidget import EtSTEDWidget
 from .FFTWidget import FFTWidget
 from .HoloWidget import HoloWidget
 from .HoliSheetWidget import HoliSheetWidget
-from .SquidStageScanWidget import SquidStageScanWidget
+from .LEDMatrixWidget import LEDMatrixWidget
+from .WellPlateWidget import WellPlateWidget
 from .FocusLockWidget import FocusLockWidget
 from .AutofocusWidget import AutofocusWidget
-from .LEDMatrixWidget import LEDMatrixWidget
 from .ImageWidget import ImageWidget
 from .LaserWidget import LaserWidget
 from .PositionerWidget import PositionerWidget
 from .RecordingWidget import RecordingWidget
 from .SLMWidget import SLMWidget
+from .UC2ConfigWidget import UC2ConfigWidget
 from .SIMWidget import SIMWidget
 from .MCTWidget import MCTWidget
+from .HistoScanWidget import HistoScanWidget
+from .PixelCalibrationWidget import PixelCalibrationWidget
+from .SquidStageScanWidget import SquidStageScanWidget
+from .ISMWidget import ISMWidget
 from .ScanWidget import ScanWidget
 from .SettingsWidget import SettingsWidget
 from .ULensesWidget import ULensesWidget
+from .WatcherWidget import WatcherWidget
 from .ViewWidget import ViewWidget
+from .TilingWidget import TilingWidget
 from .basewidgets import WidgetFactory
-from .WellPlateWidget import WellPlateWidget
